Use more rustworkx functionality for find_lines()
The `find_lines()` function was computing all the simple paths between all the pairs
of nodes in the coupling graph with a nested for loop. However, rustworkx has a
native function `all_pairs_all_simple_paths()` (documented here:
https://qiskit.org/ecosystem/rustworkx/apiref/rustworkx.all_pairs_all_simple_paths.html)
which does this more efficiently as it's all written in rust and is also multithreaded so
it will compute all the paths much more efficiently. This commit reworks the internals of
`find_lines()` function to use this function directly and improve the performance of the
graph search.